pyproject.toml
[project]
name = "mcp-weatherapi"
version = "0.1.0"
description = "MCP server for WeatherAPI.com"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.2.0",
    "httpx>=0.27.0"
]

[project.scripts]
mcp-weatherapi = "weather_server:mcp"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
